Fix German grammar: README.md
Fix German grammar in README.md as far as it does not
inhibit understanding the examples by people without
a working knowledge of German.

For that reason, we refrain from translating "title"
to "Überschrift", as "Titel" is good enough German
for this purpose and still understandable for someone
understanding English, but without a working knowledge
of German.